- Start implementing coding standards - Add first coding-standards action version
