+++ .phpcs.xml.dist
@@ -1,0 +1,15 @@
+<?xml version="1.0"?>
+<ruleset name="KuantoKusta">
+	<config name="testVersion" value="7.0-"/>
+	<rule ref="PHPCompatibilityWP"/>
+	<rule ref="WordPress">
+		<!-- Ignore the need @package tag on the file comments + full stop at the end of comments -->
+		<exclude name="Squiz.Commenting.FileComment.MissingPackageTag"/>
+		<exclude name="Squiz.Commenting.InlineComment.InvalidEndChar"/>
+		<!--<exclude name="Squiz.Commenting.FileComment.Missing"/>-->
+		<!-- No Yoda conditions! -->
+		<exclude name="WordPress.PHP.YodaConditions.NotYoda"/>
+		<!-- Query Monitor -->
+		<exclude name="WordPress.NamingConventions.ValidHookName.UseUnderscores"/>
+	</rule>
+</ruleset>
